refactor: extract HasImagePathTrait instead of redundunt method
