+++ app/Models/Alternative.php
@@ -2,13 +2,12 @@
 
 namespace App\Models;
 
+use App\Traits\HasImagePath;
 use App\Traits\HasTags;
-use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
-use Illuminate\Support\Facades\URL;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -16,6 +15,7 @@ use Spatie\MediaLibrary\MediaCollections\Models\Media;
 class Alternative extends Model implements HasMedia
 {
     use HasFactory;
+    use HasImagePath;
     use HasTags;
     use InteractsWithMedia;
 
@@ -32,39 +32,6 @@ class Alternative extends Model implements HasMedia
             'id' => 'integer',
             'approved_at' => 'timestamp',
         ];
-    }
-
-    /**
-     * Retrieves the appropriate image path:
-     * - First checks for an optimized version.
-     * - If the optimized version is not available, it checks for the original image.
-     * - If neither the optimized nor original image is available, a default image path is returned.
-     */
-    protected function imagePath(): Attribute
-    {
-        return Attribute::make(
-            get: function () {
-                $firstMedia = $this->getMedia()->first();
-
-                $optimizedPath = $firstMedia?->getPath('optimized');
-                $optimizedUrl = $firstMedia?->getUrl('optimized');
-
-                $originalPath = $firstMedia?->getPath();
-                $originalUrl = $firstMedia?->getUrl();
-
-                $defaultUrl = URL::asset('storage/images/companies/default/company.webp');
-
-                if (file_exists($optimizedPath)) {
-                    return $optimizedUrl;
-                }
-
-                if (file_exists($originalPath)) {
-                    return $originalUrl;
-                }
-
-                return $defaultUrl;
-            }
-        );
     }
 
     /**
