+++ app/Models/Company.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Enums\CompanyPersonType;
+use App\Traits\HasImagePath;
 use App\Traits\HasTags;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Casts\Attribute;
@@ -12,7 +13,6 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Support\Carbon;
-use Illuminate\Support\Facades\URL;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -22,6 +22,7 @@ use Spatie\Sluggable\SlugOptions;
 class Company extends Model implements HasMedia
 {
     use HasFactory;
+    use HasImagePath;
     use HasSlug;
     use HasTags;
     use InteractsWithMedia;
@@ -71,42 +72,6 @@ class Company extends Model implements HasMedia
         );
     }
 
-    /**
-     * Retrieves the appropriate image path:
-     * - First checks for an optimized version.
-     * - If the optimized version is not available, it checks for the original image.
-     * - If neither the optimized nor original image is available, a default image path is returned.
-     */
-    protected function imagePath(): Attribute
-    {
-        return Attribute::make(
-            get: function () {
-                $firstMedia = $this->getMedia()->first();
-
-                $optimizedPath = $firstMedia?->getPath('optimized');
-                $optimizedUrl = $firstMedia?->getUrl('optimized');
-
-                $originalPath = $firstMedia?->getPath();
-                $originalUrl = $firstMedia?->getUrl();
-
-                $defaultUrl = URL::asset('storage/images/companies/default/company.webp');
-
-                if (file_exists($optimizedPath)) {
-                    return $optimizedUrl;
-                }
-
-                if (file_exists($originalPath)) {
-                    return $originalUrl;
-                }
-
-                return $defaultUrl;
-            }
-        );
-    }
-
-    /**
-     * The attributes that should be cast to native types.
-     */
     protected function casts(): array
     {
         return [
