+++ app/Models/Person.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Traits\HasImagePath;
 use App\Traits\HasTags;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Casts\Attribute;
@@ -9,7 +10,6 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
-use Illuminate\Support\Facades\URL;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -19,6 +19,7 @@ use Spatie\Sluggable\SlugOptions;
 class Person extends Model implements HasMedia
 {
     use HasFactory;
+    use HasImagePath;
     use HasSlug;
     use HasTags;
     use InteractsWithMedia;
@@ -51,36 +52,11 @@ class Person extends Model implements HasMedia
     }
 
     /**
-     * Retrieves the appropriate image path:
-     * - First checks for an optimized version.
-     * - If the optimized version is not available, it checks for the original image.
-     * - If neither the optimized nor original image is available, a default image path is returned.
+     * Overrides the imagePath method from the HasImagePath trait
      */
     protected function imagePath(): Attribute
     {
-        return Attribute::make(
-            get: function () {
-                $firstMedia = $this->getMedia()->first();
-
-                $optimizedPath = $firstMedia?->getPath('optimized');
-                $optimizedUrl = $firstMedia?->getUrl('optimized');
-
-                $originalPath = $firstMedia?->getPath();
-                $originalUrl = $firstMedia?->getUrl();
-
-                $defaultUrl = URL::asset('storage/images/people/default/user.webp');
-
-                if (file_exists($optimizedPath)) {
-                    return $optimizedUrl;
-                }
-
-                if (file_exists($originalPath)) {
-                    return $originalUrl;
-                }
-
-                return $defaultUrl;
-            }
-        );
+        return $this->resolveImagePath('default', 'storage/images/people/default/user.webp');
     }
 
     /**
